Refactors regional observations download
Refactors/adds code for downloading regional observations for

Switzerland
================================================================================

Scripts are intended to be executed in the following sequence:

- `scripts/Switzerland_Metadata_Download.py`
  Download the ISD station metadata file and produce:
  > `data/switzerland_stations.2020-2023.txt`:
station metadata of Swiss stations for which observations are available for
download between 2020-12-03 and 2023-12-31.
- `scripts/Switzerland_Observations_Download.py`
  Download the ISD station observations and produce:
  > `data/USAF_ID-WBAN_ID-YYYY.gz`:
files with observations of Swiss stations for which observations are
available for download between 2020-12-03 and 2023-12-31.
- `scripts/Switzerland_Observations_Load.py`
  Load the ISD station observations into an xarray Dataset and produce:
  > `data/switzerland_stations.2020-2023.nc`:
netCDF file with observations of Swiss stations for which observations are
available for download between 2020-12-03 and 2023-12-31.

Contiguous US
================================================================================

Scripts are intended to be executed in the following sequence:

- `scripts/CONUS_Metadata_Download.py`
  Download the ISD station metadata file and produce:
  > `data/conus_stations.2020-2025.txt`:
station metadata of CONUS stations for which observations are available for
download between 2020-12-03 and 2025-07-31.
- `scripts/CONUS_Observations_Download.py`
  Download the ISD station observations and produce:
  > `data/USAF_ID-WBAN_ID-YYYY.gz`:
files with observations at CONUS stations for which observations are
available for download between 2020-12-03 and 2025-07-31.
- `scripts/CONUS_Observations_Load.py`
  Load the ISD station observations into an xarray Dataset and produce:
  > `data/conus_stations.2020-2025.nc`:
netCDF file with observations at CONUS stations for which observations are
available for download between 2020-12-03 and 2025-07-31.