Add 'UTC' variable to the Stations.observations xarray DataSet
When initializing a Station object from a netCDF file, a variable 'UTC' with the
UTC time as datatime objects is added to the Stations.observations xarray
DataSet for convenience.

When saving the resulting Stations.observations xarray DataSet to a netCDF file
using Stations.write_observations2netcdf, the UTC variable is skipped, as it
holds datatime objects, which cannot be saved to netCDF file. The netCDF file
will have a 'time' variable of course in CF conventions.